+++ 0x1E-search_algorithms/search_algos.h
@@ -2,7 +2,7 @@
 #define SEARCH_ALGOS
 #include <stdlib.h>
 int linear_search(int *array, size_t size, int value);
-void search_array(size_t low, size_t high);
+void search_array(int low, int high);
 int binary_search(int *array, size_t size, int value);
 
 #endif/*end header enclosement*/
